fix next-auth Session type augmentation
