+++ apps/app/tsconfig.json
@@ -8,7 +8,6 @@
 	},
 	"include": [
 		"./next-env.d.ts",
-		"../../packages/auth/next-auth.d.ts",
 		"./**/*.ts",
 		"./**/*.tsx",
 		"./**/*.js",
