--- packages/types/.eslintrc.cjs
@@ -1,4 +1,0 @@
-/* eslint-disable import/no-unused-modules */
-module.exports = {
-	extends: ['@weareinreach/eslint-config'],
-}
