--- packages/types/.lintstagedrc.cjs
@@ -1,2 +1,0 @@
-/* eslint-disable import/no-unused-modules */
-module.exports = require('@weareinreach/config/lint-staged')
