--- packages/types/.prettierignore
@@ -1,14 +1,0 @@
-node_modules
-.next
-public
-storybook-static
-dist
-
-pnpm-lock.yaml
-*.log
-
-.gitignore
-.npmignore
-.prettierignore
-.DS_Store
-.eslintignore
