--- packages/types/index.d.ts
@@ -1,4 +1,0 @@
-/* eslint-disable import/no-unused-modules */
-// codegen:start {preset: barrel, include: lib/*.d.ts}
-export * from './lib/environment.d'
-// codegen:end
