--- packages/types/lib/environment.d.ts
@@ -1,10 +1,0 @@
-/* eslint-disable import/no-unused-modules */
-/* eslint-disable @typescript-eslint/no-unused-vars */
-declare namespace NodeJS {
-	interface ProcessEnv {
-		COGNITO_ACCESS_KEY: string
-		COGNITO_SECRET: string
-		COGNITO_CLIENT_ID: string
-		COGNITO_CLIENT_SECRET: string
-	}
-}
