--- packages/types/package.json
@@ -1,23 +1,0 @@
-{
-	"name": "@weareinreach/types",
-	"version": "0.100.0",
-	"private": true,
-	"license": "GPL-3.0-only",
-	"type": "module",
-	"main": "./index.d.ts",
-	"types": "./index.d.ts",
-	"scripts": {
-		"clean:node": "rm -rf ./node_modules/ || true",
-		"format": "prettier --cache --cache-strategy metadata --write --ignore-unknown .",
-		"preinstall": "npx only-allow pnpm",
-		"lint": "eslint --cache .",
-		"lint:fix": "eslint --cache --fix ."
-	},
-	"devDependencies": {
-		"@types/node": "16",
-		"@weareinreach/config": "0.100.0",
-		"@weareinreach/eslint-config": "0.100.0",
-		"eslint": "8.32.0",
-		"typescript": "4.9.4"
-	}
-}
