--- packages/types/tsconfig.json
@@ -1,3 +1,0 @@
-{
-	"extends": "@weareinreach/config/tsconfig/base.json"
-}
